+++ rs/src/complete/image.rs
@@ -1,5 +1,6 @@
 use crate::complete::base::skip;
 use nom::number::complete::{be_u16 as parse_be_u16, be_u32 as parse_be_u32};
+use nom::IResult as NomResult;
 
 pub struct ImageDimensions {
   pub width: usize,
@@ -37,68 +38,116 @@ pub fn get_png_image_dimensions(input: &[u8]) -> Result<ImageDimensions, ()> {
   })
 }
 
-pub fn get_jpeg_image_dimensions(input: &[u8]) -> Result<ImageDimensions, ()> {
-  let mut dimensions: Option<ImageDimensions> = None;
-
-  for chunk in read_jpeg_chunks(input) {
-    let code: u8 = chunk[1];
-    // SOF: 0b110000xx
-    if (code & 0xfc) == 0xc0 && chunk.len() >= 9 {
-      let frame_height: u16 = ((chunk[5] as u16) << 8) + (chunk[6] as u16);
-      let frame_width: u16 = ((chunk[7] as u16) << 8) + (chunk[8] as u16);
-      dimensions = match dimensions {
-        None => Some(ImageDimensions {
-          width: frame_width as usize,
-          height: frame_height as usize,
-        }),
-        d => d,
-      };
+pub fn get_jpeg_image_dimensions(mut input: &[u8]) -> Result<ImageDimensions, ()> {
+  loop {
+    let (next_input, chunk) = take_next_jpeg_chunk(input).map_err(|_| ())?;
+    if chunk.is_some() {
+      // Assert progress
+      debug_assert!(next_input.len() < input.len());
     }
-  }
-
-  match dimensions {
-    Some(d) => Ok(d),
-    None => Err(()),
+    input = next_input;
+    if let Some(chunk) = chunk {
+      if chunk.len() < 9 {
+        continue;
+      }
+      let code: u8 = chunk[1];
+      if !is_jpeg_sof(code) {
+        continue;
+      }
+      // At this point we have an SOFn chunk with at least 9 bytes
+      let frame_height = u16::from_be_bytes([chunk[5], chunk[6]]);
+      let frame_width = u16::from_be_bytes([chunk[7], chunk[8]]);
+      // There may be multiple SOF chunks, we return the dimension corresponding
+      // to the first one.
+      return Ok(ImageDimensions {
+        width: usize::from(frame_width),
+        height: usize::from(frame_height),
+      });
+    } else {
+      // End of chunks
+      return Err(());
+    }
   }
 }
 
-fn read_jpeg_chunks(input: &[u8]) -> Vec<&[u8]> {
-  let mut next_chunk_start: Option<&[u8]> = find_next_chunk(input);
+/// Finds the next jpeg chunk (or marker/segment/sequence)
+///
+/// JPEG files are organized as a stream of chunks.
+/// A chunk starts with a marker: the two-byte sequence `[0xff, marker_code]` where
+/// `0x00 < marker_code < 0xff`. (Consecutive `0xff` values represent padding and `[0xff, 0x00]` is
+/// an escaped `0x00` value).
+/// Garbage is allowed between chunks, so you have to scan the input to find the marker signaling
+/// the start of the chunk.
+/// There are two types of chunks:
+/// - Standalone markers: they consist in only the marker
+/// - Marker sequences: the marker is followed by a sequence. The sequence starts with a
+/// `sequence_size` field followed by data. The `sequence_size` is a big-endian U16, it includes the
+/// `sequence_size` field its and data, but not the sequence marker.
+///
+/// This functions returns the next JPEG chunk.
+/// If a chunk is found, preceding garbage is skipped and the result is `Ok(suffix, Some(chunk))`
+/// where `chunk` is the whole chunk (marker and optional sequence) and `suffix` the remaining
+/// input.
+/// If a chunk is not found, it returns `Ok(&[], None)` (the input is consumed).
+/// If an error occurs, it returns `Err`. Errors can occur if a sequence marker is found but there
+/// is not enough data to read the sequence.
+fn take_next_jpeg_chunk(input: &[u8]) -> NomResult<&[u8], Option<&[u8]>> {
+  use nom::bytes::complete::take;
+  use nom::combinator::map;
 
-  let mut result: Vec<&[u8]> = Vec::new();
-
-  while let Some(chunk) = next_chunk_start {
-    let code: u8 = chunk[1];
-    let mut size: usize = 2;
-    if (code >= 0xc0 && code <= 0xc7)
-      || (code >= 0xc9 && code <= 0xcf)
-      || (code >= 0xda && code <= 0xef)
-      || code == 0xfe
-    {
-      size += ((chunk[2] as usize) << 8) + (chunk[3] as usize);
-    }
-    next_chunk_start = find_next_chunk(&chunk[size..]);
-    match &next_chunk_start {
-      Some(next) => result.push(&chunk[..(chunk.len() - next.len())]),
-      None => result.push(chunk),
-    }
-  }
-
-  result
-}
-
-fn find_next_chunk(input: &[u8]) -> Option<&[u8]> {
   let mut search: usize = 0;
   while search + 1 < input.len() {
     let cur_byte: u8 = input[search];
-    let next_byte: u8 = input[search + 1];
-    if cur_byte == 0xff && (next_byte != 0x00 && next_byte != 0xff) {
-      return Some(&input[search..]);
+    let marker_type: Option<JpegMarkerType> = if cur_byte == 0xff {
+      get_jpeg_marker_type(input[search + 1])
     } else {
-      search += 1;
+      None
+    };
+    match marker_type {
+      Some(marker_type) => {
+        // Consume the padding bytes.
+        // It won't panic because we have `search + 1 < input.len()`
+        let input = &input[search..];
+        return match marker_type {
+          JpegMarkerType::Standalone => map(take(2usize), Some)(input),
+          JpegMarkerType::Sequence => {
+            let (_, size) = parse_be_u16(input)?;
+            map(take(2usize + usize::from(size)), Some)(input)
+          }
+        };
+      }
+      None => search += 1,
     }
   }
-  None
+  // Reached end of input without finding a marker
+  Ok((&[], None))
+}
+
+#[derive(Debug, Eq, PartialEq)]
+enum JpegMarkerType {
+  Standalone,
+  Sequence,
+}
+
+/// Returns the JPEG marker type (standalone or sequence) for `code`.
+///
+/// If the code does not correspond to a marker type, returns `None`.
+fn get_jpeg_marker_type(code: u8) -> Option<JpegMarkerType> {
+  match code {
+    0x00 => None,                                    // Escaped `0x00`
+    0x01 => Some(JpegMarkerType::Standalone),        // TEM (Temporary)
+    0xd0..=0xd7 => Some(JpegMarkerType::Standalone), // RSTn (Reset)
+    0xd8 => Some(JpegMarkerType::Standalone),        // SOI (Start of image)
+    0xd9 => Some(JpegMarkerType::Standalone),        // EOI (End of image)
+    0xff => None,                                    // Padding
+    _ => Some(JpegMarkerType::Sequence),             // Reserved (non-standalone) or sequence marker
+  }
+}
+
+/// Checks if the provided code corresponds to a Start of frame (SOFn) JPEG marker
+fn is_jpeg_sof(code: u8) -> bool {
+  // SOFn: 0b110000xx
+  code & 0xfc == 0xc0
 }
 
 pub fn get_gif_image_dimensions(input: &[u8]) -> Result<ImageDimensions, ()> {